The panel is a header, a picture, a strip and some tags
graph-item-details-panel.tsx was 821 lines and four unrelated subjects. Split
along the seams the redesign left in it.

THE MONITOR is the one part of a panel that is not about a clip's data at all:
a video element, a low-res twin for scrubbing, a canvas holding the outgoing
frame across a cut, and the seek arithmetic keeping the three in agreement. It
takes its own machinery with it — `useSeekedVideo` and `useFrameCrossfade`
exist to serve that one element, so leaving them in the parent would mean
threading four refs and a proxy source through a prop list to reach the only
thing that uses them.

THE TRIM STRIP is everything about DURATION: the whole source as a filmstrip,
the showing window with its grips, the typed in/out points. It takes the width
measurement with it, which existed only to size the strip.

Two things fell out on the way. `restingFrame` is gone — every panel rests on
its own first frame now, so a prop that is always the same value is one a
caller cannot get wrong, and the seek arithmetic loses a branch that only ever
ran for the other value. And `previewTime` moved next to its only caller.

821 → 458 lines, with a 338-line monitor and a 161-line strip either side of
it. Behaviour-preserving by construction; the modal's 22 story interactions
pass unchanged, which is the only claim worth making about a split.
